Implement initial resolve behaviour on feed.
Implement a test set of resolving behaviour on feed.  Mainly, this resolve
the regular feed content so that it matches what was last fetched.  It
doesn't deal with items yet (deleted or not).  Also, add a custom deep
equal function (copied from reflect) that ignores certain feeds.  This is
used to ignore the Model field given from the riak library, as it differs,
but not in a way a test cares about.