Stop removing the allow multi setting on buckets.
When the test suite is run against a real cluster (invovling multiple nodes),
the changing of the allow multi setting races the gossip protocol to make
the change known.  This causes some expected conflicts to not occur, and thus
break the suite now and again.  Thus just don't do it.  Unfortunately, that
means the cluster isn't left in a pristine condition when the suite is finished.