Parallelize data fetching for item comparison.
For large data sets, it takes very little time to compare an item, but a
large set of time to fetch it.  Thus use go routines to fetch many items
simultaneously, and compare them as they finish.  Note that most items
seem to be required before the first few fetch (by random chance), so memory
requirements are increased and it takes some time before a comparison may
happen.