Make sure that the fetched time is also stored in the database.
Since the fetched time is actually important (especially for resolving conflicts),
make sure it is actually stored in the database.  Also for the RSS parser test,
just use the fetched time from the json files to calculate the nextchecktime,
which means a custom compare is no longer necessary, since the json files can
just have the correct next check time in them.