Maintain a separate map of inserted feed items instead of linearly searching.
Since double dealt with items should be ignored, the db update would linearly
search for duplicates through all previously inserted items.  However, this
caused a huge slow down when dealing with a large number of to insert items.
Thus use a map instead storing the state.  This might end up being a bad idea
since to use a map, the keys need to be converted to strings, which involves
a copy of data, and thus more garbage (yay!).