Implement an initial fetcher for rss feeds that actually runs.
Implement an initial fetcher executable that will fetch feeds live from the
internet.  It uses a separate package to maintain a list of feeds it should
ensure are fetched, but it will fetch all feeds that exist in the database.
It currently connects to just a development cluster, and the feed list is
easily customized by someone else for their own use.