Fix test suite to work with latest github.com/gorilla/feeds.
The feeds package no longer outputs seconds into the generated RSS feeds.
Due to this, seconds can no longer be checked.  To fix for now, simple kill
any seconds in the test data that is used for RSS parsing.  This means the
seconds are no longer tested, but it shouldn't be too much of an issue.